+++ packages/bookmark/src/BookmarkListRoot.tsx
@@ -22,6 +22,7 @@ interface BookmarkListRootProps {
 const BookmarkListRoot: React.FC<BookmarkListRootProps> = ({ bookmarks, doc, store, onJumpToDest }) => {
     const containerRef = React.useRef<HTMLDivElement>();
     const bookmarkElementsRef = React.useRef<HTMLLIElement[]>([]);
+    const [currentFocused, setCurrentFocused] = React.useState(0);
 
     const updateLinkAnnotation = (bookmark: PdfJs.Outline, links: Record<string, HTMLElement>): void => {
         const dest = bookmark.dest;
@@ -51,6 +52,55 @@ const BookmarkListRoot: React.FC<BookmarkListRootProps> = ({ bookmarks, doc, sto
         });
     };
 
+    const handleKeyDown = (e: React.KeyboardEvent) => {
+        switch (e.key) {
+            case 'ArrowDown':
+                activateNextItem();
+                break;
+
+            case 'ArrowUp':
+                activatePreviousItem();
+                break;
+
+            default:
+                break;
+        }
+    };
+
+    const activateNextItem = () => {
+        const container = containerRef.current;
+        if (!container) {
+            return;
+        }
+
+        const items = bookmarkElementsRef.current;
+        const nextItem = currentFocused + 1;
+
+        if (nextItem < items.length) {
+            if (currentFocused >= 0) {
+                items[currentFocused].setAttribute('tabindex', '-1');
+            }
+            setCurrentFocused(nextItem);
+        }
+    };
+
+    const activatePreviousItem = () => {
+        const container = containerRef.current;
+        if (!container) {
+            return;
+        }
+
+        const items = bookmarkElementsRef.current;
+        const prevItem = currentFocused - 1;
+
+        if (prevItem >= 0) {
+            if (currentFocused >= 0) {
+                items[currentFocused].setAttribute('tabindex', '-1');
+            }
+            setCurrentFocused(prevItem);
+        }
+    };
+
     React.useEffect(() => {
         store.subscribe('linkAnnotations', handleLinkAnnotationsChanged);
 
@@ -77,9 +127,27 @@ const BookmarkListRoot: React.FC<BookmarkListRootProps> = ({ bookmarks, doc, sto
         }
     }, []);
 
+    React.useEffect(() => {
+        const bookmarkElements = bookmarkElementsRef.current;
+        if (bookmarkElements.length === 0 || currentFocused < 0 || currentFocused > bookmarkElements.length) {
+            return;
+        }
+
+        const bookmarkEle = bookmarkElements[currentFocused];
+        bookmarkEle.setAttribute('tabindex', '0');
+        bookmarkEle.focus();
+    }, [currentFocused]);
+
     return (
-        <div ref={containerRef}>
-            <BookmarkList bookmarks={bookmarks} depth={0} doc={doc} isRoot={true} store={store} onJumpToDest={jumpToDest} />
+        <div ref={containerRef} onKeyDown={handleKeyDown}>
+            <BookmarkList
+                bookmarks={bookmarks}
+                depth={0}
+                doc={doc}
+                isRoot={true}
+                store={store}
+                onJumpToDest={jumpToDest}
+            />
         </div>
     );
 };